Arduino Controlled Light Dimmer
Testing output parameters. Implementation in process. 

Value starts at 0, with the press of the button, value rises while the button is pressed until max value. 
When the button is released, value stays fixed. If the button is pressed again, value decreases while the button is pressed until min value.


To do: when first pressed, value has to be set to max. 
Optional: if the button is pressed for a small amount of time while the output is not 0, set output to 0.