Imrpoved Button Function
If powered off, when the button is pressed output will go to MAX state, no matter how long is pressed.
If button is pressed for shorter than preset period when the output is different than 0, output is set to 0.

Dimming functionality. If button is pressed and is not in 0 state, value will go up or down. Direction depends on a previous state.